refactor: keep files in index
